fix raycasting, add touch support, add optimized shader version

+++ Assets/Scripts/BrushProjector.cs
@@ -19,7 +19,7 @@ public class BrushProjector : MonoBehaviour {
 
     void Start() {
         if(paintedRT == null){
-            paintedRT = new RenderTexture(512, 512, 0, RenderTextureFormat.RG32);
+            paintedRT = new RenderTexture(512, 512, 0, RenderTextureFormat.RG16);
             paintedRT.enableRandomWrite = true;
         }
